Update MPModManager.lua to prevent mod disable
I deleted line 103. It prevents mods being disabled when launching BeamNG through the BeamMP launcher and joining a server. Many people have asked on Reddit for the past couple years asking if there was a way to stop disabling of mods so they don't have to enable UI mods again, or so friends can play in Vanilla servers with all the mods they use and it syncs perfectly, as well as adding mods to servers that don't have all the mods you want. In my case it's also to save space so you don't have to re-download mods you already have for a server and so I don't have to enable a script in the editor and then re activate my 82 mods with my friend every single time we play, takes like 20 mins, and prevented us from actually playing the game. Now we have been playing non stop and love it! I want to share the love with others! I think this would work best as an on and off switch so you can keep the current functionality or use this new method of mods not disabling. Has been a GAME CHANGER for me and have had no issues! Hope you guys enjoy!